requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
pyyaml==6.0.1